+++ src/main/java/br/ufsc/lapesd/riefederator/federation/cardinality/impl/ThresholdCardinalityComparator.java
@@ -34,26 +34,26 @@ public class ThresholdCardinalityComparator implements CardinalityComparator {
     @Override
     public int compare(@Nonnull Cardinality l, @Nonnull Cardinality r) {
         if (l.getReliability().equals(r.getReliability()))
-            return Integer.compare(l.getValue(0), r.getValue(0));
+            return Long.compare(l.getValue(0), r.getValue(0));
         if (l.getReliability() == UNSUPPORTED || l.getReliability() == NON_EMPTY)
-            return Integer.compare(huge, r.getValue(huge-1));
+            return Long.compare(huge, r.getValue(huge-1));
         if (r.getReliability() == UNSUPPORTED || r.getReliability() == NON_EMPTY)
-            return Integer.compare(l.getValue(huge-1), huge);
+            return Long.compare(l.getValue(huge-1), huge);
 
         if ((l.getReliability() == GUESS || l.getReliability() == LOWER_BOUND) &&
             (r.getReliability() == GUESS || r.getReliability() == LOWER_BOUND)) {
-            return Integer.compare(l.getValue(0), r.getValue(0));
+            return Long.compare(l.getValue(0), r.getValue(0));
         }
         if ((l.getReliability() == UPPER_BOUND || l.getReliability() == EXACT) &&
             (r.getReliability() == UPPER_BOUND || r.getReliability() == EXACT)) {
-            return Integer.compare(l.getValue(0), r.getValue(0));
+            return Long.compare(l.getValue(0), r.getValue(0));
         }
         if (l.getReliability() == UPPER_BOUND || l.getReliability() == EXACT)
-            return Integer.compare(l.getValue(0), r.getValue(0)+large);
+            return Long.compare(l.getValue(0), r.getValue(0)+large);
         if (r.getReliability() == UPPER_BOUND || r.getReliability() == EXACT)
-            return Integer.compare(l.getValue(0)+large, r.getValue(0));
+            return Long.compare(l.getValue(0)+large, r.getValue(0));
         assert false;
 
-        return Integer.compare(l.getValue(0), r.getValue(0));
+        return Long.compare(l.getValue(0), r.getValue(0));
     }
 }
